Music_stylesheet.css
This addition is a stylesheet for the website, currently contains just the CSS for the Nav bar as well as the main body background colour. Make sure to clone the repository before updating the repository in case issues occur and update on what has been changed. I will back up frequently locally in case of accidential deletion.